Chore: Rename directory from parser to src
