+++ src/interface.ts
